Card #255 unread count performance improvements fix for data test properties and enable sql logging

+++ root/api/src/main/java/org/patientview/api/controller/ConversationController.java
@@ -75,7 +75,7 @@ public class ConversationController extends BaseController<ConversationControlle
     @RequestMapping(value = "/user/{userId}/conversations/unreadcount", method = RequestMethod.GET,
             produces = MediaType.APPLICATION_JSON_VALUE)
     @ResponseBody
-    public ResponseEntity<Integer> getUnreadConversationCount(@PathVariable("userId") Long userId)
+    public ResponseEntity<Long> getUnreadConversationCount(@PathVariable("userId") Long userId)
             throws ResourceNotFoundException {
         return new ResponseEntity<>(conversationService.getUnreadConversationCount(userId), HttpStatus.OK);
     }
